refactor(basket): Phase 5 extract event-consumer slices
- Features/OrderCreated and Features/ProductPriceUpdated each hold a
  sealed IEventHandler plus an AddXxxSlice() DI extension that wires
  the shared AddEventHandler<TEvent,THandler>.
- Program.cs chains the two slice extensions; the inline AddEventHandler
  calls and the now-empty IntegrationEvents/ folder are removed.
