+++ basket-microservice/Basket.Service/Features/OrderCreated/OrderCreatedSliceExtensions.cs
@@ -1,0 +1,13 @@
+using Basket.Service.Contracts.Integration;
+using ECommerce.Shared.Infrastructure.EventBus;
+
+namespace Basket.Service.Features.OrderCreated;
+
+internal static class OrderCreatedSliceExtensions
+{
+    public static IServiceCollection AddOrderCreatedSlice(this IServiceCollection services)
+    {
+        services.AddEventHandler<OrderCreatedEvent, OrderCreatedHandler>();
+        return services;
+    }
+}
